refactor: split gym state module
