+++ world.py
@@ -1,6 +1,5 @@
-"""World state facade; player state lives in :mod:`world.player`."""
+"""World state facade backed by per-domain state modules."""
 import importlib
-import json
 import math
 import os
 import sys
@@ -13,6 +12,7 @@ import settings as _cfg
 __path__ = [os.path.join(os.path.dirname(__file__), "world")]
 player = importlib.import_module(__name__ + ".player")
 spawns = importlib.import_module(__name__ + ".spawns")
+gyms = importlib.import_module(__name__ + ".gyms")
 
 from world.player import (
     BADGE_DEFINITIONS, DEFAULT_AVATAR, ITEM_GREAT_BALL, ITEM_POKE_BALL,
@@ -21,8 +21,9 @@ from world.player import (
     _players, _record_type_badges, _safe_name, _unpack_badge_thresholds,
     account_names, accounts, acting_as, add_candy, add_coins, add_item,
     add_stardust, avatar_for, bag_count, bag_full, bag_items, candy,
-    check_login, codename_for, current, drain_badge_pending, has_password,
-    level_bounds, level_for_xp, new_uid, onboarding_needed, record_badge_progress,
+    check_login, codename_for, current, drain_badge_pending, get_caught,
+    has_password, level_bounds, level_for_xp, new_uid, onboarding_needed,
+    record_badge_progress,
     room_in_bag, save, set_password, spend, spend_coins, take_item, team_for,
     type_badges_from_game_master, use,
 )
@@ -34,11 +35,17 @@ from world.spawns import (
     remember_spawn, remove_spawn,
 )
 
+from world.gyms import (
+    BATTLES, GYMS, GYMS_FILE, RAID, RAID_FILE, _defender_coins,
+    _defender_minutes, _max_defenders, _raid_member, clear_gym,
+    collect_gym_returns, deploy, gym_guard, gym_members, gym_team,
+    is_deployed, is_raid_uid, load_gyms, load_raid, my_team, raid, recall,
+    save_gyms, save_raid, set_raid, set_team, time_left,
+)
+
 HERE = player.HERE
 SAVE_FILE = player.SAVE_FILE
 SAVES_DIR = player.SAVES_DIR
-GYMS_FILE = os.path.join(HERE, "gyms.json")
-RAID_FILE = os.path.join(HERE, "raid.json")
 _TYPE_BADGES = player._TYPE_BADGES
 
 
@@ -145,12 +152,6 @@ def caught():
         return list(current().CAUGHT)
 
 
-def get_caught(uid):
-    with _lock:
-        for c in current().CAUGHT:
-            if c["uid"] == uid:
-                return dict(c)
-    return None
 
 
 def update_caught(uid, **fields):
@@ -480,285 +481,6 @@ def berry_mult(encounter_id, consume=False):
     return m
 
 
-# ==============================================================================
-# Shared World State Data & Spawns
-# ==============================================================================
-
-GYMS = {}                          # fort_id -> [{uid, pokemon_id, cp, trainer, team}]
-BATTLES = {}                       # battle_id -> live battle state
-RAID = {"on": False, "pokemon_id": 150, "cp": 3000, "trainer": "raid"}
-
-
-# ==============================================================================
-# Shared Raids & Bonus Spawns
-# ==============================================================================
-
-def load_raid():
-    global RAID
-    try:
-        with open(RAID_FILE, "r", encoding="utf-8") as fh:
-            d = json.load(fh)
-        RAID.update(
-            {
-                "on": bool(d.get("on", False)),
-                "pokemon_id": int(d.get("pokemon_id", 150) or 150),
-                "cp": int(d.get("cp", 3000) or 3000),
-                "trainer": str(d.get("trainer", "raid") or "raid"),
-            }
-        )
-    except (OSError, ValueError, TypeError):
-        pass
-    return dict(RAID)
-
-
-def save_raid():
-    try:
-        tmp = RAID_FILE + ".tmp"
-        with open(tmp, "w", encoding="utf-8") as fh:
-            json.dump(RAID, fh, indent=1)
-        os.replace(tmp, RAID_FILE)
-    except OSError:
-        pass
-
-
-def set_raid(on=None, pokemon_id=None, cp=None, trainer=None):
-    """Turn raid mode on/off."""
-    sent_home = 0
-    with _lock:
-        if on is not None:
-            RAID["on"] = bool(on)
-        if pokemon_id is not None:
-            RAID["pokemon_id"] = max(1, min(151, int(pokemon_id)))
-        if cp is not None:
-            RAID["cp"] = max(10, min(9999, int(cp)))
-        if trainer is not None:
-            RAID["trainer"] = str(trainer)[:16] or "raid"
-        if RAID["on"]:
-            sent_home = sum(len(v) for v in GYMS.values())
-            GYMS.clear()
-    save_gyms()
-    save_raid()
-    return dict(RAID), sent_home
-
-
-def raid():
-    with _lock:
-        return dict(RAID)
-
-
-def _raid_member(fort_id):
-    uid = (abs(hash(("raid", fort_id))) & 0x3FFFFFFFFFFFFFFF) | 1
-    return {
-        "uid": uid,
-        "pokemon_id": int(RAID["pokemon_id"]),
-        "cp": int(RAID["cp"]),
-        "trainer": RAID.get("trainer", "raid"),
-        "team": 0,
-        "raid": True,
-        "deployed_ms": int(time.time() * 1000),
-    }
-
-
-def is_raid_uid(fort_id, uid):
-    return RAID["on"] and _raid_member(fort_id)["uid"] == uid
-
-
-
-
-# ==============================================================================
-# Shared Gyms Operations
-# ==============================================================================
-
-def _defender_minutes():
-    return _cfg.get("gyms", "defender_minutes", env="DEFENDER_MINUTES", cast=float)
-
-
-def _defender_coins():
-    return _cfg.get("gyms", "defender_coins", env="DEFENDER_COINS", cast=int)
-
-
-def _max_defenders():
-    return _cfg.get("gyms", "max_defenders", cast=int)
-
-
-def save_gyms():
-    try:
-        tmp = GYMS_FILE + ".tmp"
-        with open(tmp, "w", encoding="utf-8") as fh:
-            json.dump({"gyms": GYMS}, fh, indent=1)
-        os.replace(tmp, GYMS_FILE)
-    except OSError:
-        pass
-
-
-def load_gyms():
-    try:
-        with open(GYMS_FILE, "r", encoding="utf-8") as fh:
-            d = json.load(fh)
-        GYMS.clear()
-        GYMS.update(
-            {k: v for k, v in (d.get("gyms") or {}).items() if isinstance(v, list)}
-        )
-        return True
-    except (OSError, ValueError):
-        pass
-    try:                       # migrate gyms out of old single-player save
-        with open(SAVE_FILE, "r", encoding="utf-8") as fh:
-            d = json.load(fh)
-        GYMS.update(
-            {k: v for k, v in (d.get("gyms") or {}).items() if isinstance(v, list)}
-        )
-        if GYMS:
-            save_gyms()
-    except (OSError, ValueError):
-        pass
-    return False
-
-
-def my_team():
-    """The team this trainer picked in game."""
-    t = current().TEAM
-    return t if t else _cfg.get("gyms", "team", env="TEAM", cast=int)
-
-
-def set_team(team):
-    """SetPlayerTeam."""
-    p = current()
-    with _lock:
-        if p.TEAM:
-            return 2, p.TEAM
-        p.TEAM = max(1, min(3, int(team)))
-        out = p.TEAM
-    p.save()
-    return 1, out
-
-
-def gym_members(fort_id):
-    with _lock:
-        if RAID["on"]:
-            return [_raid_member(fort_id)]
-        return list(GYMS.get(fort_id, []))
-
-
-def gym_team(fort_id):
-    """Which team holds this gym (0 = neutral/white)."""
-    ms = gym_members(fort_id)
-    return ms[0].get("team", 0) if ms else 0
-
-
-def is_deployed(uid):
-    with _lock:
-        return any(m["uid"] == uid for ms in GYMS.values() for m in ms)
-
-
-def deploy(fort_id, uid, trainer=None, team=None):
-    p = current()
-    trainer = trainer or p.username
-    team = team if team is not None else my_team()
-    c = get_caught(uid)
-    if not c:
-        return False, "unknown pokemon"
-    with _lock:
-        members = GYMS.setdefault(fort_id, [])
-        if members and members[0].get("team", team) != team:
-            return False, "held by another team"
-        if any(m["uid"] == uid for m in members):
-            return False, "already deployed"
-        if len(members) >= _max_defenders():
-            return False, "gym full"
-        members.append(
-            {
-                "uid": uid,
-                "pokemon_id": c["pokemon_id"],
-                "cp": c["cp"],
-                "trainer": trainer,
-                "team": team,
-                "owner": p.username,
-                "deployed_ms": int(time.time() * 1000),
-            }
-        )
-    save_gyms()
-    return True, "ok"
-
-
-def recall(fort_id, uid):
-    with _lock:
-        members = GYMS.get(fort_id, [])
-        n = len(members)
-        GYMS[fort_id] = [m for m in members if m["uid"] != uid]
-        changed = n != len(GYMS[fort_id])
-        if not GYMS[fort_id]:
-            GYMS.pop(fort_id, None)
-    save_gyms()
-    return changed
-
-
-def clear_gym(fort_id):
-    """Every defender is knocked out -- the gym goes neutral."""
-    with _lock:
-        GYMS.pop(fort_id, None)
-    save_gyms()
-
-
-def gym_guard(fort_id):
-    ms = gym_members(fort_id)
-    if not ms:
-        return None
-    best = max(ms, key=lambda m: m.get("cp", 0))
-    return best["pokemon_id"], best["cp"], 500 * len(ms)
-
-
-def collect_gym_returns():
-    """Bring home any Pokemon that has served its shift and pay its coins."""
-    now = int(time.time() * 1000)
-    cutoff = _defender_minutes() * 60_000
-    coins = _defender_coins()
-    faint = _cfg.get("pokemon", "faint_after_gym", cast=bool)
-    returned = []
-    with _lock:
-        for fid in list(GYMS):
-            keep = []
-            for m in GYMS.get(fid, []):
-                dep = m.get("deployed_ms")
-                if dep is None:
-                    m["deployed_ms"] = now
-                    keep.append(m)
-                elif now - dep >= cutoff:
-                    returned.append(
-                        (fid, m["pokemon_id"], coins, m.get("owner"), m["uid"])
-                    )
-                else:
-                    keep.append(m)
-            if keep:
-                GYMS[fid] = keep
-            else:
-                GYMS.pop(fid, None)
-    mine = []
-    if returned:
-        save_gyms()
-        me = current()
-        for fid, pid, c, owner, uid in returned:
-            if owner and owner != me.username:
-                continue
-            with _lock:
-                me.COINS += c
-                if faint:
-                    for pk in me.CAUGHT:
-                        if pk["uid"] == uid:
-                            pk["stamina"] = 0
-                            break
-            mine.append((fid, pid, c))
-        if mine:
-            me.save()
-    return mine
-
-
-def time_left(fort_id, uid):
-    for m in gym_members(fort_id):
-        if m["uid"] == uid:
-            dep = m.get("deployed_ms") or 0
-            return max(0, int(_defender_minutes() * 60 - (time.time() - dep / 1000)))
-    return 0
 
 
 # ==============================================================================
@@ -779,7 +501,7 @@ class _WorldFacade(types.ModuleType):
         "bag_count", "bag_full", "room_in_bag", "add_coins", "spend_coins",
         "add_stardust", "candy", "add_candy", "spend", "avatar_for",
         "codename_for", "team_for", "new_uid", "level_for_xp", "level_bounds",
-        "type_badges_from_game_master",
+        "get_caught", "type_badges_from_game_master",
     }
     _player_dynamic = {
         "BAG", "CAUGHT", "CANDY", "STARDUST", "XP", "LEVEL", "COINS",
@@ -787,6 +509,11 @@ class _WorldFacade(types.ModuleType):
         "BERRIES", "APPLIED", "MAX_POKEMON", "MAX_ITEMS", "CLAIMED_LEVELS",
         "STATS", "BADGE_PROGRESS", "BADGE_LEVELS", "BADGE_PENDING",
     }
+    _spawn_exports = {
+        "SPAWNS", "DESPAWNED", "FORT_MODIFIERS", "BONUS_SPAWNS", "_MAX_SPAWNS",
+    }
+    _gym_exports = {"GYMS", "BATTLES", "RAID", "GYMS_FILE", "RAID_FILE"}
+
 
     def __getattr__(self, name):
         if name in self._player_dynamic:
@@ -796,6 +523,12 @@ class _WorldFacade(types.ModuleType):
     def __setattr__(self, name, value):
         if name in self._player_exports:
             setattr(player, name, value)
+        if name == "SAVE_FILE":
+            gyms.SAVE_FILE = value
+        if name in self._spawn_exports:
+            setattr(spawns, name, value)
+        if name in self._gym_exports:
+            setattr(gyms, name, value)
         super().__setattr__(name, value)
